- PoliceReports will be ranked based on Cosinus Similarity

+++ src/main/java/de/htw/berlin/f4/ai/suchmaschinenpolizeiberichte/model/policeReport/RankedPoliceReport.java
@@ -9,5 +9,5 @@ import lombok.*;
 @Builder
 public class RankedPoliceReport {
     private String idToOrigin;
-    private int score;
+    private double score;
 }
